refactor: move files into correct folder and add PreviousNextButton.js to AddExistingContext.js
