+++ pages/explore/manage-room/AddExistingItem.js
